Refactor: 링크 검색 기능 필터링 로직 분리, 타입 추가
