+++ src/types/link.ts
@@ -1,0 +1,8 @@
+export interface LinkData {
+  id: number;
+  title: string;
+  description: string;
+  created_at: string;
+  image_source?: string;
+  url: string;
+}
